// src/models/User.ts
export type User = {
    lastName: string;
    firstName: string;
    email: string;
    phone: string;
    password: string;
    businessName: string;
    town: string;
    address: string;
    linkedin: string;
    twitter: string;
    facebook: string;
    instagram: string;
    whatsapp: string
    picture: string;
}